requirements.txt
absl-py==0.7.1
astor==0.8.0
audioread==2.1.7
decorator==4.4.0
gast==0.2.2
grpcio==1.21.1
h5py==2.9.0
joblib==0.13.2
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
librosa==0.6.3
llvmlite==0.28.0
Markdown==3.1.1
mock==3.0.5
numba==0.43.1
numpy==1.16.3
Pillow==6.0.0
protobuf==3.7.1
python-Levenshtein==0.12.0
resampy==0.2.1
scikit-learn==0.21.2
scipy==1.3.0
six==1.12.0
tensorboard==1.13.1
tensorboardX==1.7
tensorflow-estimator==1.13.0
termcolor==1.1.0
torch==1.1.0
torchaudio==0.2
tqdm==4.32.1
warprnnt-pytorch==0.1
Werkzeug==0.15.4
wget==3.2
